update: 后端，友链权限精细控制
还需要进行数据库迁移

+++ src/backend/NanTingBlog.API/Controllers/FriendslinkController.cs
@@ -1,10 +1,8 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using NanTingBlog.API.Dtos.Blogs;
 using NanTingBlog.API.Services.Blog;
-#if RELEASE
-using Microsoft.AspNetCore.Authorization;
 using NanTingBlog.API.Services.Identitys;
-#endif
 
 namespace NanTingBlog.API.Controllers;
 
@@ -21,9 +19,10 @@ public class FriendslinkController(FriendslinkService service) : ControllerBase
     [HttpGet("getall")]
     public ActionResult<BaseResult<List<Friendslink>>> GetAll()
     {
+        var allFriendlink = service.GetAll();
         var result = new BaseResult<List<Friendslink>>()
         {
-            Data = service.GetAll()
+            Data = allFriendlink
         };
         return Ok(result);
     }
@@ -31,9 +30,7 @@ public class FriendslinkController(FriendslinkService service) : ControllerBase
     /// <summary>
     /// 删除给定友链
     /// </summary>
-#if RELEASE
-    [Authorize(Policy = PolicyTypes.ADMIN)]
-#endif
+    [Authorize(Policy = PolicyTypes.USER)]
     [HttpPost("delete")]
     public async Task<ActionResult<BaseResult<string>>> DeleteById([FromBody] DeleteByIdInput input)
     {
@@ -43,28 +40,83 @@ public class FriendslinkController(FriendslinkService service) : ControllerBase
             result.Data = "无效的id";
             return Ok(result);
         }
-        await service.DeleteByKeyAsync(input.Id);
-        return Ok(result);
+        return await Divide(
+            admin: async userid => {
+                await service.DeleteByKeyAsync(input.Id);
+                return Ok(result);
+            }, 
+            user: async userId => {
+                if (await IsArticleMine(userId, input.Id)) {
+                    await service.DeleteByKeyAsync(input.Id);
+                    return Ok(result);
+                }
+                return Ok(BaseResult<string>.CreateError("身份错误"));
+            }
+        );
     }
 
     /// <summary>
     /// <br> 添加或更新友链，如果要添加，请不要传入Id </br>
     /// <br> 无论如何都不要传入创建时间和创建字串 </br>
     /// </summary>
-#if RELEASE
-    [Authorize(Policy = PolicyTypes.ADMIN)]
-#endif
+    [Authorize(Policy = PolicyTypes.USER)]
     [HttpPost("addOrUpdate")]
     public async Task<ActionResult<BaseResult<string>>> AddOrUpdate([FromBody] Friendslink newFriendslink)
     {
-        var result = new BaseResult<string>();
+        var result = BaseResult<string>.CreateError("无效操作");
         if (newFriendslink == null) {
-            result.Code = 500;
-            result.Data = "无效的新链";
             return Ok(result);
         }
-        await service.UpdateOrAddAsync(newFriendslink);
-        return Ok(result);
+        return await Divide(
+            admin: async userId => {
+                await service.UpdateOrAddAsync(newFriendslink);
+                return Ok(BaseResult<string>.Create("成功"));
+            },
+            user: async userId => {
+                if (userId == null) {
+                    return Ok(BaseResult<string>.CreateError("请先登录"));
+                }
+                newFriendslink.State = STATES.Pending;
+                newFriendslink.UserId = userId;
+                await service.UpdateOrAddAsync(newFriendslink);
+                return Ok(BaseResult<string>.Create("成功"));
+            }, BaseResult<string>.CreateError("身份错误")
+        );
+    }
+
+    private async Task<ActionResult<BaseResult<string>>> Divide(
+        Func<string?, Task<ActionResult<BaseResult<string>>>> admin, 
+        Func<string?, Task<ActionResult<BaseResult<string>>>> user,
+        BaseResult<string>? defaultRet = null
+        )
+    {
+        var userId = HttpContext.GetUserId();
+        if (HttpContext.IsAdmin()) {
+            return await admin(userId);
+        } else if (HttpContext.IsUser()) {
+            return await user(userId);
+        }
+        defaultRet ??= BaseResult<string>.CreateError("无效身份");
+        return Ok(defaultRet);
+    }
+
+    /// <summary>
+    /// 这个文章是此用户的吗？
+    /// </summary>
+    /// <returns>文章为null，用户id为null，文章用户id和用户id不相同 均返回false</returns>
+    private async Task<bool> IsArticleMine(string? userId, string fid)
+    {
+        var targetFl = await service.QueryByKeyAsync(fid);
+        if (targetFl == null) {
+            return false;
+        }
+        if (userId == null) {
+            return false;
+        }
+        if (userId != targetFl.UserId) {
+            return false;
+        }
+        return true;
     }
 
 #pragma warning disable CS1591 // 缺少对公共可见类型或成员的 XML 注释
